Don't always sort; allow explicit arrays.
Rather than call ranges, markers, and measures, see if they're a function first.
Sorting the markers is unhelpful: they won't occlude each other and it's helpful
to be able to style the first marker more prominently.
Sorting the measures is unhelpful for the same reason, we want to always know s0
is the actual value and s1 is the forecast. However, for rendering, we sort them
to prevent occlusion.